oauth crs: noop restructure
